refactor(ui): Move hud ui logic to new `ui::hud` mod
